requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
mcp>=1.0
